requirements.txt
accelerate==1.13.0
basicsr==1.4.2
diffusers==0.35.2
facexlib==0.3.0
gfpgan==1.3.8
numpy==1.26.4
opencv-python==4.11.0.86
Pillow==9.5.0
realesrgan==0.3.0
requests==2.32.3
safetensors==0.5.3
simple-lama-inpainting==0.1.2
tqdm==4.67.1
transformers==4.52.4
